FAT: (incomplete, buggy) Implement writing. FAT: Implement FAT table parsing. TODO: Make read_sync use the FAT cluster chains. Vfs: A lot of bug fixes related to page caches, reading, and writing. klog.c: Fix file header to say oboskrnl/klog.c instead of klog.h Fat: Rename driver to slowfat, as a parody of 'fastfat', Window's FAT driver. NOTE: This commit is very buggy, and it is likely to have a lot of code that will be changed drastically in upcoming commits.
