+++ CMakeLists.txt
@@ -138,7 +138,7 @@ if (OBOS_ARCHITECTURE STREQUAL "x86_64")
 	endif()
 add_subdirectory("src/drivers/generic/initrd")
 add_subdirectory("src/drivers/generic/ahci")
-add_subdirectory("src/drivers/generic/fat")
+add_subdirectory("src/drivers/generic/slowfat")
 add_subdirectory("src/drivers/test_driver")
 add_subdirectory("src/isogen")
 if (OBOS_ARCHITECTURE STREQUAL "m68k")
