+++ config/hyper.cfg
@@ -5,7 +5,7 @@ protocol=ultra
 higher-half-exclusive=true
 kernel-as-module=true
 video-mode=auto
-cmdline='-initrd-module=initrd -initrd-driver-module=initrd_driver --root-fs-partid initrd --working-set-cap=8388608 --load-modules=/fat'
+cmdline='-initrd-module=initrd -initrd-driver-module=initrd_driver --root-fs-partid initrd --working-set-cap=8388608 --load-modules=/slowfat'
 
 binary:
 	path="/obos/oboskrnl"
