+++ scripts/generate_initrd-x86_64.sh
@@ -6,7 +6,7 @@ fi
 cd $(git rev-parse --show-toplevel)
 cp out/uart tar
 cp out/ahci tar
-cp out/fat tar
+cp out/slowfat tar
 if [[ ! -d tar/dev ]]
 then
     mkdir tar/dev
